fixed compilation on msvc

+++ psycle/src/psycle/host/Configuration.hpp
@@ -137,7 +137,7 @@ namespace psycle
 				public:
 					midi_type() : groups_(16), velocity_(0x0c), pitch_(1), raw_()
 					{
-						for(std::size_t i(0) ; i < groups().size() ; ++i) group(i).message() = group(i).command() = i + 1;
+						for(std::size_t i(0) ; i < groups().size() ; ++i) group(i).message() = group(i).command() = static_cast<int>(i + 1);
 					}
 
 				public:
